Make WeightedJobsIdle a double, not an int #7313

+++ src/condor_schedd.V6/scheduler.h
@@ -156,8 +156,8 @@ struct LiveJobCounters {
 struct SubmitterCounters {
   int JobsRunning;
   int JobsIdle;
-  int WeightedJobsRunning;
-  int WeightedJobsIdle;
+  double WeightedJobsRunning;
+  double WeightedJobsIdle;
   int JobsHeld;
   int JobsFlocked;
   int JobsFlockedHere; // volatile field use to hold the JobsRunning calculation when sending submitter adds to flock collectors
@@ -917,8 +917,8 @@ private:
 	void			initLocalStarterDir( void );
 	void	noShadowForJob( shadow_rec* srec, NoShadowFailure_t why );
 	bool			jobExitCode( PROC_ID job_id, int exit_code );
-	int			calcSlotWeight(match_rec *mrec);
-	int			guessJobSlotWeight(JobQueueJob * job);
+	double			calcSlotWeight(match_rec *mrec);
+	double			guessJobSlotWeight(JobQueueJob * job);
 	
 	// -----------------------------------------------
 	// CronTab Jobs
